fix incompatibility with postgres/adminer

+++ database/migrations/2025_06_01_144400_create_ranking_wrestler_averages_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->foreignId('ranking_id')->constrained()->cascadeOnDelete();
             $table->foreignId('wrestler_id')->constrained()->cascadeOnDelete();
             $table->unsignedInteger('votes_count')->default(0);
-            $table->unsignedInteger('votes_sum')->default(0);
-            $table->decimal('average_vote', 5, 2)->default(0);
+            $table->decimal('votes_sum', 8, 2)->default(0);
+            $table->decimal('average_vote', 8, 2)->default(0);
             $table->timestamps();
 
             $table->unique(['ranking_id', 'wrestler_id']);
